+++ packages/vite-config/src/node/plugins/vconsole.ts
@@ -1,6 +1,5 @@
-import { isTest } from 'std-env'
 import { viteVConsole, type viteVConsoleOptions } from 'vite-plugin-vconsole'
 
 export function vConsole(options: viteVConsoleOptions) {
-  return isTest ? viteVConsole(options) : undefined
+  return viteVConsole(options)
 }